Conflicts:
	bundles/at.medevit.elexis.emediplan.core/META-INF/MANIFEST.MF
	bundles/at.medevit.elexis.emediplan.ui/META-INF/MANIFEST.MF
	bundles/at.medevit.elexis.emediplan.ui/src/at/medevit/elexis/emediplan/ui/handler/CreateAndOpenHandler.java
	bundles/ch.elexis.base.ch.arzttarife/plugin.xml
	bundles/ch.elexis.base.ch.labortarif_2009/src/ch/elexis/base/ch/labortarif_2009/ui/Labor2009Selector.java
	bundles/ch.elexis.omnivore.ui/META-INF/MANIFEST.MF
	bundles/ch.elexis.omnivore.ui/build.properties
	bundles/ch.itmed.lmz.risch.laborder/META-INF/MANIFEST.MF
	tests/at.medevit.elexis.emediplan.core.test/EMediplanTests.launch
	tests/at.medevit.elexis.emediplan.core.test/src/at/medevit/elexis/emediplan/core/internal/EMediplanServiceImplTest.java

+++ bundles/at.medevit.elexis.emediplan.core/src/at/medevit/elexis/emediplan/core/model/chmed16a/Medicament.java
@@ -37,7 +37,7 @@ public class Medicament {
 	public String Roa;
 	public int Rep;
 	public int Subs;
-	public float NbPack;
+	public int NbPack;
 	public List<PrivateField> PFields;
 	public transient IArtikelstammItem artikelstammItem;
 	public transient String dosis;
